add PatternController, started Enemy/PlayerActors
The PatternController allows for a set pattern of movements for a CharacterActor to be put into place, and allows for looping as well. Started the Enemy/Player actor classes, to allow for a bit of distinction between different kinds of actors. Kinda walking around in the dark here, not sure what I'm doing, but seems to be going ok.