add basic onAction triggering
Pressing Z next to an NPC will now display the word ACTION on the screen, plus a unique number for the NPC. works by following the outline i laid out a few commits ago.. first by eliminating actors based on their XY pos and the direction the player is facing, and then sorting the remaining actors by distance from the player. then, chooses the closest actor within range and uses its onAction interaction, if it has one. if it doesnt, it will move to the next closest actor in range. range is determined as sqrt(physWidth + physHeight) for actors