create input queue
the InputController class now contains a queue of keyDown events. With this, GameScene can now handle inputs in a queue, rather than as a state. this solves the problem of holding down Z constantly re-triggering the message interaction with NPCs, making the message never show up. now the user can hold Z and see the message just fine, but also spam Z to see the message re-draw over and over again.