draw message window background
created the window package, and now have an object to represent a message window. draws the background for a message, then displays the message. need to modify GameScene to stop all animation, and allow for the window to receive inputs... idling when the message is finished, and afterwards closing when the action key is hit