add message input controls
pressing button0 during a message will skip the message scrolling to the end, and pressing again will end the message window OR go to the next window in the message. pressing button1 during a message will cancel the message window and close it, prematurely or no. also added functionality for message interactions to split up a string into a format appropriate for a message window. need to start using a better font, one with a fixed width preferrably.