refactor GameScene to use Managers
Using Managers as a sort of a cascading state machine. Managers have act and button press methods, that return a boolean if the act or button press should end with that manager. ie if a window manager has windows, it will handle a button press, preventing button presses from cascading down to the actor manager. likewise, if a window manager is set to halting, it will return true to say it has handled the act method, and will prevent the actormanager class from acting. it might be prudent to separate the actormanager into two or three classes, because the player actor management seems to be especially specific compared to regular NPC management, and it might be worthwhile to separate NPCs from Enemies from Environment actors, etc.. but may not require separate manager classes for each one. All in all, the cascading model allows for cleaner code in the Scene class, and separates the concern of handling actors, windows, and all their specific behaviors from one another. For now I have left the Map management in the Scene class, as that seems to be appropriate.