optimize draw functions
Actors and map tiles will now only be drawn if they are within the game screen. this required adding some methods to the actor class to specify the total width and height of the actor (which needs to be different per actor class, eg with the MassiveActor class there are a lot of sprites to draw). the drawing of map tiles was moved to the GameMap class, giving GameMap a draw method. Also started experimenting with colors a little. To consider: need to implement some sort of way to specify a source for every interaction, so that the interaction can also relay commands back to the source. for instance, talking to an NPC will send a message box to the scene, but itd be nice to have the NPC face the actor when the message window opens... to do that, we need the actor from which the Interaction originated.