make strides towards refactoring input movement
need to make movement from input agnostic of the actual input. made some progress towards this tonight by heavily refactoring the input controller (now renamed InputState, relinquishing the duties of controller to the GameScene, as it should be), and adding the ability for Managers to handle directional inputs (sorta). for now, it's all more or less a foundation... none of the code written tonight actually does anything different, other than sate my need to code