fix draw priority for scene, provide unique actor names
draw priority refers to the order sprites and tiles are drawn in order to maintain the illusion of depth.. this is done by drawing tiles and sprites row by row, starting from the top of the screen going downwards. unique actor names are given to actors in order to be able to refer to them by name later on, for key interactions etc.